+++ CHANGELOG.md
@@ -7,6 +7,10 @@
 - `vector_matching` check for finding queries with incorrect `on()` or `ignoring()`
   keywords.
 
+### Fixed
+
+- `comparison` check would trigger false positive for rules using `unless` keyword.
+
 ## [v0.1.2]
 
 ### Fixed